+++ librairies/Security/Sentry.php
@@ -286,7 +286,16 @@ class Sentry extends SystemComponent
 
         if ($sauvegarder)
         {
-            setcookie('ladecadanse_remember', $cookie, time() + 1209600, '/'); // 2 semaines
+            $cookieOptions = [
+                'expires' => strtotime('+15 days'),
+                'path' => '/',
+                //'domain' => '.example.com', // leading dot for compatibility or use subdomain
+                'secure' => true,
+                'httponly' => true,
+                'samesite' => 'Lax'
+            ];
+
+            setcookie('ladecadanse_remember', $cookie, $cookieOptions);
         }
 
         /* 	echo "Veuillez patienter...";
@@ -347,7 +356,7 @@ class Sentry extends SystemComponent
             //setcookie('ladecadanse_remember', '', time() - 3600); // semble ne pas fonctionner
             unset($_COOKIE['ladecadanse_remember']);
 
-            setcookie('ladecadanse_remember', null, -1, '/');
+            setcookie('ladecadanse_remember', null, 1);
         }
     }